requirements.txt
asttokens==2.4.1
attrs==24.2.0
azure-ai-ml==1.22.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-compute==33.0.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
azure-search-documents==11.6.0b3
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
azure-storage-file-share==12.20.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
huggingface-hub==0.26.2
idna==3.10
ipython==8.29.0
ipywidgets==8.1.5
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
portalocker==2.10.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydash==8.0.4
Pygments==2.18.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
six==1.16.0
stack-data==0.6.3
strictyaml==1.7.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
